Changed the Factory to a builder and moved it to utils, created both a stream for all events and for the upcoming events. Improved the readability and added comments in relevant places
